# apps/api-java/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "ChatSpace API",
    "description" : "Slack風チャットアプリ ChatSpace のREST API。詳細はdocs/要件定義書.mdを参照。",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/workspaces/{workspaceId}/dms/{dmId}/read" : {
      "post" : {
        "tags" : [ "dm-controller" ],
        "operationId" : "markRead",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/channels/{channelId}/read" : {
      "post" : {
        "tags" : [ "channel-controller" ],
        "operationId" : "markRead_1",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/notifications/{notificationId}/read" : {
      "post" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "markRead_2",
        "parameters" : [ {
          "name" : "notificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/notifications/read-all" : {
      "post" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "markAllRead",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/notifications" : {
      "get" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "list_5",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "unreadOnly",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "cursorCreatedAt",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "cursorId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NotificationListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Cursor" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "NotificationListResponse" : {
        "type" : "object",
        "properties" : {
          "notifications" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NotificationResponse"
            }
          },
          "nextCursor" : {
            "$ref" : "#/components/schemas/Cursor"
          }
        }
      },
      "NotificationResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "MENTION", "DM", "CHANNEL_INVITE", "WORKSPACE_INVITE", "THREAD_REPLY" ]
          },
          "workspaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "channelId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "dmId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "messageId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "threadParentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "fromUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "text" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "readAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    }
  }
}